# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Motors Connect API - Driver",
    "description": "Documentação completa da API para o aplicativo de motoristas do Motors Connect.",
    "version": "1.0.0",
    "contact": {
      "name": "Suporte Motors Connect",
      "email": "logistica.robertmarinho@gmail.com"
    }
  },
  "servers": [
    {
      "url": "https://motors-connect-api.vercel.app",
      "description": "Servidor de Produção"
    },
    {
      "url": "http://localhost:3000",
      "description": "Servidor Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Login": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string" }
        }
      },
...
    }
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["Autenticação"],
        "summary": "Login de usuário",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/Login" }
            }
          }
        },
        "responses": {
          "200": { "description": "Login bem-sucedido" },
          "401": { "description": "Credenciais inválidas" }
        }
      }
    },
...
  }
}